feat: move schedule logic from recorder package to main
